Updates on Geo Eff calculation
