+++ app/runGeoEffFDEvtSim.cpp
@@ -40,46 +40,57 @@ using namespace std;
 // Include customized functions and constants
 #include "Helpers.h"
 
-int main(){
-
+// The program which do all translations and rotations when moving event from FD to ND:
+// 0. FD: read event from FD MC ntuple: before earth curvature rotation
+// 1. FD to ND: after earth curvature rotation
+// 2. ND: move back to the beam center
+// 3. ND to ND: translate from OnAxis to OffAxis
+// 4. ND: get after eigen rotated vectors for step 3
+// 5. ND: generate random throws
+int main()
+{
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
   //
-  // Branches to be read from n-tuple produced from FD MC
+  // 0. FD: read event from FD MC ntuple: before earth curvature rotation
   //
-
-  int FD_Run;
-  int FD_SubRun;
-  int FD_Event;
-  int FD_Sim_nNumu;
-  double FD_Gen_numu_E;
-  int FD_Sim_nMu;
-  int FD_CCNC_truth; // Choose CCNC
-  int FD_neuPDG; // neutrino PDG
-  double FD_Sim_mu_start_vx; // unit: cm?
-  double FD_Sim_mu_start_vy;
-  double FD_Sim_mu_start_vz;
-  double FD_Sim_mu_end_vx;
-  double FD_Sim_mu_end_vy;
-  double FD_Sim_mu_end_vz;
-  double FD_Sim_mu_start_px;
-  double FD_Sim_mu_start_py;
-  double FD_Sim_mu_start_pz;
-  double FD_Sim_mu_start_E;
-  double FD_Sim_mu_end_px;
-  double FD_Sim_mu_end_py;
-  double FD_Sim_mu_end_pz;
-  double FD_Sim_mu_end_E;
-  double FD_Sim_hadronic_Edep_a2;
-  int FD_Sim_n_hadronic_Edep_a;
-  vector<float> *FD_Sim_hadronic_hit_Edep_a2 = 0; // Need initialize 0 here to avoid error
-  vector<float> *FD_Sim_hadronic_hit_x_a     = 0; // Same as mu pos unit: cm?
-  vector<float> *FD_Sim_hadronic_hit_y_a     = 0;
-  vector<float> *FD_Sim_hadronic_hit_z_a     = 0;
-
-  // Read ntuple from FD MC
   TChain *t = new TChain("MyEnergyAnalysis/MyTree");
-  // ntuple path on FNAL dunegpvm machine
-  t->Add("/home/fyguo/FDEff/srcs/myntuples/myntuples/MyEnergyAnalysis/myntuple.root");
+  // Ntuple path on FNAL dunegpvm machine
+  // For FNAL machine:
+  t->Add("/dune/app/users/flynnguo/FDEff/srcs/myntuples/myntuples/MyEnergyAnalysis/myntuple.root");
 
+  // Define variables for FD event
+  int FD_Run; // # of the run being processed
+  int FD_SubRun; // # of the sub-run being processed
+  int FD_Event; // # of the event being processed
+  int FD_Sim_nNumu; // # of Sim muon neutrinos (numu and numubar)
+  double FD_Gen_numu_E; // Energy of generator level neutrino [GeV]
+  int FD_Sim_nMu; // # of Sim muons (mu+/mu-)
+  int FD_CCNC_truth; // 0 =CC 1 =NC
+  int FD_neuPDG; // Generator level neutrino PDG
+  double FD_Sim_mu_start_vx; // Position of the muon trajectory at start point on the x-axis [cm]
+  double FD_Sim_mu_start_vy; // Position of the muon trajectory at start point on the y-axis [cm]
+  double FD_Sim_mu_start_vz; // Position of the muon trajectory at start point on the z-axis [cm]
+  double FD_Sim_mu_end_vx; // Position of the muon trajectory at end point on the x-axis [cm]
+  double FD_Sim_mu_end_vy; // Position of the muon trajectory at end point on the y-axis [cm]
+  double FD_Sim_mu_end_vz; // Position of the muon trajectory at end point on the z-axis [cm]
+  double FD_Sim_mu_start_px; // Momentum of the muon trajectory at start point on the x-axis [GeV]
+  double FD_Sim_mu_start_py; // Momentum of the muon trajectory at start point on the y-axis [GeV]
+  double FD_Sim_mu_start_pz; // Momentum of the muon trajectory at start point on the z-axis [GeV]
+  double FD_Sim_mu_start_E; // Energy of leading mu at start point [GeV]
+  double FD_Sim_mu_end_px; // Momentum of the muon trajectory at end point on the x-axis [GeV]
+  double FD_Sim_mu_end_py; // Momentum of the muon trajectory at end point on the y-axis [GeV]
+  double FD_Sim_mu_end_pz; // Momentum of the muon trajectory at end point on the z-axis [GeV]
+  double FD_Sim_mu_end_E; // Energy of leading mu at end point [GeV]
+  double FD_Sim_hadronic_Edep_a2; // Total amount of energy released by ionizations in the event (from Geant4 simulation) [MeV]
+  int FD_Sim_n_hadronic_Edep_a; // # of hadronic energy deposits
+  vector<float> *FD_Sim_hadronic_hit_Edep_a2 = 0; // Need initialize 0 here to avoid error
+  vector<float> *FD_Sim_hadronic_hit_x_a = 0; // Position of each energy deposit on the x-axis [cm]
+  vector<float> *FD_Sim_hadronic_hit_y_a = 0; // Position of each energy deposit on the y-axis [cm]
+  vector<float> *FD_Sim_hadronic_hit_z_a = 0; // Position of each energy deposit on the z-axis [cm]
+
+  // Extract event info from ntuple
   t->SetBranchAddress("Run",                      &FD_Run);
   t->SetBranchAddress("SubRun",                   &FD_SubRun);
   t->SetBranchAddress("Event",                    &FD_Event);
@@ -108,84 +119,32 @@ int main(){
   t->SetBranchAddress("Sim_hadronic_hit_x_a",     &FD_Sim_hadronic_hit_x_a);
   t->SetBranchAddress("Sim_hadronic_hit_y_a",     &FD_Sim_hadronic_hit_y_a);
   t->SetBranchAddress("Sim_hadronic_hit_z_a",     &FD_Sim_hadronic_hit_z_a);
-
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
   //
   // Declare variables used in this program
   //
+  int nentries = 0; // Total input events
+  float vetoEnergyFD; // Total hadron deposited energy in FD veto region
+  int iwritten = 0; // Output event counter
+  float decayZbeamCoord; // Decay point (neutrino production point) in beam coordinate [cm]
+  float decayXdetCoord; // Decay point (neutrino production point) in detector coordinate on the x-axis [cm]
+  float decayYdetCoord; // Decay point (neutrino production point) in detector coordinate on the y-axis [cm]
+  float decayZdetCoord; // Decay point (neutrino production point) in detector coordinate on the z-axis [cm]
+  vector<float> HadronHitEdeps; // Hadron hit segment energy deposits [MeV]
+  vector<float> HadronHitPoss;  // Hadron hit segment energy deposits position [cm]
 
-  int nentries = 0;        // Total input events
-  float vetoEnergyFD;      // Total hadron deposited energy in FD veto region
-  float vetoEnergyFD_new;  // Add vetoEnergy for new restriction
-  int iwritten = 0;        // Output event counter
-  float decayZbeamCoord;
-  float decayXdetCoord;
-  float decayYdetCoord;
-  float decayZdetCoord;
-  vector<float> HadronHitEdeps;
-  vector<float> HadronHitPoss;
-
-  vector<double> ND_off_axis_pos_vec; // unit: meters, ND off-axis choices for each FD evt: 1st element is randomized for each evt
-  vector<double> ND_vtx_vx_vec;          // unit: cm, vtx x choices for each FD evt in ND volume: 1st element is randomized for each evt
-  int ND_off_axis_pos_steps = 0;
-  int vtx_vx_steps = 0;
-
-  // Initialize first element as -999, to be replaced by a random off-axis nd pos in each evt below
-  ND_off_axis_pos_vec.emplace_back(-999.);
-
-  if ( ND_off_axis_pos_stepsize > 0 && ND_off_axis_pos_stepsize <= OffAxisPoints[13] ) {
-    ND_off_axis_pos_steps = ( OffAxisPoints[13] - OffAxisPoints[0] ) / ND_off_axis_pos_stepsize;
-  }
-  else std::cout << "Error: please set the ND_off_axis_pos_stepsize above 0 and below max element of OffAxisPoints." << std::endl;
-
-  if (verbose) std::cout << "ND_off_axis_pos_steps: " << ND_off_axis_pos_steps << std::endl;
-
-  // The rest elements follow fixed increments from min ND local x
-  for ( int i_ND_off_axis_pos_step = 0; i_ND_off_axis_pos_step < ND_off_axis_pos_steps + 1; i_ND_off_axis_pos_step++ ){
-    ND_off_axis_pos_vec.emplace_back( i_ND_off_axis_pos_step*ND_off_axis_pos_stepsize + OffAxisPoints[0] );
-  }
-
-  if (verbose) std::cout << "ND_off_axis_pos_vec size: "<< ND_off_axis_pos_vec.size() << std::endl;
-
-  // Initialize first element as -999, to be replaced by a random vtx x in each evt below
-  ND_vtx_vx_vec.emplace_back(-999.);
-
-  if ( ND_local_x_stepsize > 0 && ND_local_x_stepsize <= ND_local_x_max ) {
-    vtx_vx_steps = ( ND_local_x_max - ND_local_x_min ) / ND_local_x_stepsize;
-  }
-  else std::cout << "Error: please set the ND_local_x_stepsize above 0 and below ND_local_x_max." << std::endl;
-
-  if (verbose) std::cout << "vtx_vx_steps: " << vtx_vx_steps << std::endl;
-
-  // The rest elements follow fixed increments from min ND local x
-  for ( int i_vtx_vx_step = 0; i_vtx_vx_step < vtx_vx_steps + 1; i_vtx_vx_step++ ){
-    ND_vtx_vx_vec.emplace_back( i_vtx_vx_step*ND_local_x_stepsize + ND_local_x_min );
-  }
-
-  if (verbose) std::cout << "ND_vtx_vx_vec size: "<< ND_vtx_vx_vec.size() << std::endl;
-
-  // Lepton info: expressed in ND coordinate sys, do not confuse with branches read above in FD coordinate sys
-  double ND_Gen_numu_E;
-  // Add veto E with two different restrictions
-  double ND_vetoEnergyFD;
-  double ND_vetoEnergyFD_new;
-
-  vector<double> ND_Sim_mu_start_vx; // Vector corresponds to randomized/stepwise vtx x
-  vector<double> ND_Sim_mu_end_vx;
-  double ND_Sim_mu_start_vy;         // Do not use float!
-  double ND_Sim_mu_start_vz;
-  double ND_Sim_mu_end_vy;
-  double ND_Sim_mu_end_vz;
-  double ND_Sim_mu_start_px;
-  double ND_Sim_mu_start_py;
-  double ND_Sim_mu_start_pz;
-  double ND_Sim_mu_start_E;
-  double ND_Sim_mu_end_px;
-  double ND_Sim_mu_end_py;
-  double ND_Sim_mu_end_pz;
-  double ND_Sim_mu_end_E;
-  // Tot hadron E dep
-  double ND_Sim_hadronic_Edep_a2;
+  vector<double> ND_off_axis_pos_vec = {0,-7*100,-30*100}; // ND off-axis positions' choices for each FD evt [cm]
+  vector<double> ND_vtx_vx_vec={-200,0,200};          // Vtx x choices for each FD evt in ND volume [cm]
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //
   // Result of hadron containment is stored in nested vectors, need to generate dictionary
+  gSystem->Exec("rm -f AutoDict*vector*vector*float*"); // Remove old dictionary if exists
   gSystem->Exec("rm -f AutoDict*vector*vector*bool*"); // Remove old dictionary if exists
   gSystem->Exec("rm -f AutoDict*vector*vector*vector*bool*");
   gSystem->Exec("rm -f AutoDict*vector*vector*vector*vector*bool*");
@@ -193,57 +152,142 @@ int main(){
   gSystem->Exec("rm -f AutoDict*vector*vector*vector*vector*uint64_t*");
   gSystem->Exec("rm -f AutoDict*vector*vector*vector*vector*vector*uint64_t*");
   // Generate new dictionary
+  gInterpreter->GenerateDictionary("vector<vector<float> >", "vector");
   gInterpreter->GenerateDictionary("vector<vector<bool> >", "vector");
   gInterpreter->GenerateDictionary("vector<vector<vector<bool> > >", "vector");
   gInterpreter->GenerateDictionary("vector<vector<vector<vector<bool> > > >", "vector");
   gInterpreter->GenerateDictionary("vector<vector<vector<uint64_t> > >", "vector");
   gInterpreter->GenerateDictionary("vector<vector<vector<vector<uint64_t> > > >", "vector");
   gInterpreter->GenerateDictionary("vector<vector<vector<vector<vector<uint64_t> > > > >", "vector");
-  // Nested vector (vetoSize > vetoEnergy) returned from function getHadronContainmentOrigin
-  vector<vector<bool> > hadron_contain_result_before_throw;
-  vector<vector<vector<bool> > > hadron_contain_result_before_throw_vec_for_vtx_vx; // One more vector: randomized/stepwise evt vtx x
-  vector<vector<vector<vector<bool> > > > ND_Sim_hadron_contain_result_before_throw; // ................ ................... ND off axis position x
   // Nested vector (vetoSize > vetoEnergy > 64_bit_throw_result) returned from function getHadronContainmentThrows
   vector<vector<vector<uint64_t> > > hadron_throw_result;
   vector<vector<vector<vector<uint64_t> > > > hadron_throw_result_vec_for_vtx_vx;   // One more vector: randomized/stepwise evt vtx x
   vector<vector<vector<vector<vector<uint64_t> > > > > ND_Sim_hadron_throw_result;   // ................ ................... ND off axis position x
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //
+  // 1. FD to ND: after earth curvature rotation
+  //
+  // Lepton info: expressed in ND coordinate sys, do not confuse with branches read above in FD coordinate sys
+  double ND_Gen_numu_E; // Energy of generator level neutrino [GeV]
+  // Muon info
+  // array: (x,y,z)<->dim=(0,1,2)
+  double ND_RandomVtx_Sim_mu_start_v[3]; // Position of the muon trajectory at start point [cm]
+  double ND_RandomVtx_Sim_mu_end_v[3]; // Position of the muon trajectory at end point [cm]
+  double ND_RandomVtx_Sim_mu_start_p[3]; // Momentum of the muon trajectory at start point on the x-axis [GeV]
+  // double ND_RandomVtx_Sim_mu_start_E; // Energy of leading mu at start point [GeV]
+  // double ND_RandomVtx_Sim_mu_end_E; // Energy of leading mu at end point [GeV]
+  // Hadron info
+  double ND_Sim_hadronic_Edep_a2; // Total amount of energy released by ionizations in the event (from Geant4 simulation) [MeV]
+  vector<vector<float>> ND_RandomVtx_Sim_hadronic_hit; // Position of each energy deposit [cm]
+  vector<float> ND_RandomVtx_Sim_hadronic_hit_xyz; // Position of each energy deposit [cm]
+  // Add veto E with two different restrictions
+  double ND_vetoEnergyFD; // Total hadron deposited energy in FD veto region
+  // Momentum conservation check
+  float ND_RandomVtx_Sim_mu_start_p_total; // Total momentum of ND_OnAxis_Sim_mu_start_p_total
 
   //
-  // A tree to store lepton info (for NN training)
-  // and result of hadron containment after applying transformations
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
   //
+  // 2. ND: move back to the beam center
+  //
+  // Muon info
+  // array: (x,y,z)<->dim=(0,1,2)
+  double ND_OnAxis_Sim_mu_start_v[3]; // Position of the muon trajectory at start point [cm]
+  double ND_OnAxis_Sim_mu_end_v[3]; // Position of the muon trajectory at end point [cm]
+  double ND_OnAxis_Sim_mu_start_p[3]; // Momentum of the muon trajectory at start point on the x-axis [GeV]
+  // Hadron info
+  vector<vector<float>> ND_OnAxis_Sim_hadronic_hit; // Position of each energy deposit [cm]
+  vector<float> ND_OnAxis_Sim_hadronic_hit_xyz;
 
+  // Momentum conservation check
+  float ND_OnAxis_Sim_mu_start_p_total; // Total momentum of ND_OnAxis_Sim_mu_start_p_total
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //
+  // 3. ND to ND: translate from OnAxis to OffAxis
+  //
+  // Muon info
+  // array: (x,y,z)<->dim=(0,1,2)
+  double ND_OffAxis_Unrotated_Sim_mu_start_v[3]; // Position of the muon trajectory at start point [cm]
+  double ND_OffAxis_Unrotated_Sim_mu_end_v[3]; // Position of the muon trajectory at end point [cm]
+  double ND_OffAxis_Unrotated_Sim_mu_start_p[3]; // Momentum of the muon trajectory at start point on the x-axis [GeV]
+  // Hadron info
+  vector<vector<float>> ND_OffAxis_Unrotated_Sim_hadronic_hit; // Position of each energy deposit [cm]
+  vector<float> ND_OffAxis_Unrotated_Sim_hadronic_hit_xyz; // Position of each energy deposit [cm]
+  // Momentum conservation check
+  float ND_OffAxis_Unrotated_Sim_mu_start_p_total; // Total momentum of ND_OffAxis_Unrotated_Sim_mu_start_p
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //
+  // 4. ND: get after eigen rotated vectors for step 4
+  //
+  // Muon info
+  // array: (x,y,z)<->dim=(0,1,2)
+  double ND_OffAxis_Sim_mu_start_v[3]; // Position of the muon trajectory at start point [cm]
+  double ND_OffAxis_Sim_mu_end_v[3]; // Position of the muon trajectory at end point [cm]
+  double ND_OffAxis_Sim_mu_start_p[3]; // Momentum of the muon trajectory at start point on the x-axis [GeV]
+  double ND_OffAxis_Sim_mu_start_E; // Energy of the muon trajectory [GeV]
+
+  // Hadron info
+  vector<vector<float>> ND_OffAxis_Sim_hadronic_hit; // Position of each energy deposit [cm]
+  vector<float> ND_OffAxis_Sim_hadronic_hit_xyz;
+  // Momentum conservation check
+  float ND_OffAxis_Sim_mu_start_p_total; // Total momentum of ND_OffAxis_Sim_mu_start_p
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //
+  // Store variables into a tree
   TTree * effTreeFD = new TTree("effTreeFD", "FD eff Tree");
   effTreeFD->Branch("ND_Gen_numu_E",                             &ND_Gen_numu_E,            "ND_Gen_numu_E/D");
-  // Add two vetoE branches
   effTreeFD->Branch("ND_vetoEnergyFD",                           &ND_vetoEnergyFD,          "ND_vetoEnergyFD/D");
-  effTreeFD->Branch("ND_vetoEnergyFD_new",                       &ND_vetoEnergyFD_new,      "ND_vetoEnergyFD_new/D");
-  effTreeFD->Branch("ND_off_axis_pos_vec",                       &ND_off_axis_pos_vec);                             // vector<double>: entries = written evts * ND_off_axis_pos_steps
-  effTreeFD->Branch("ND_Sim_mu_start_vx",                        &ND_Sim_mu_start_vx);                              // ............... entries = written evts * vtx_vx_steps, equivalent to ND_vtx_vx_vec
-  effTreeFD->Branch("ND_Sim_mu_start_vy",                        &ND_Sim_mu_start_vy,       "ND_Sim_mu_start_vy/D");   // entries = written evts
-  effTreeFD->Branch("ND_Sim_mu_start_vz",                        &ND_Sim_mu_start_vz,       "ND_Sim_mu_start_vz/D");
-  effTreeFD->Branch("ND_Sim_mu_end_vx",                          &ND_Sim_mu_end_vx);
-  effTreeFD->Branch("ND_Sim_mu_end_vy",                          &ND_Sim_mu_end_vy,         "ND_Sim_mu_end_vy/D");
-  effTreeFD->Branch("ND_Sim_mu_end_vz",                          &ND_Sim_mu_end_vz,         "ND_Sim_mu_end_vz/D");
-  effTreeFD->Branch("ND_Sim_mu_start_px",                        &ND_Sim_mu_start_px,       "ND_Sim_mu_start_px/D");
-  effTreeFD->Branch("ND_Sim_mu_start_py",                        &ND_Sim_mu_start_py,       "ND_Sim_mu_start_py/D");
-  effTreeFD->Branch("ND_Sim_mu_start_pz",                        &ND_Sim_mu_start_pz,       "ND_Sim_mu_start_pz/D");
-  effTreeFD->Branch("ND_Sim_mu_start_E",                         &ND_Sim_mu_start_E,        "ND_Sim_mu_start_E/D");
-  effTreeFD->Branch("ND_Sim_mu_end_px",                          &ND_Sim_mu_end_px,         "ND_Sim_mu_end_px/D");
-  effTreeFD->Branch("ND_Sim_mu_end_py",                          &ND_Sim_mu_end_py,         "ND_Sim_mu_end_py/D");
-  effTreeFD->Branch("ND_Sim_mu_end_pz",                          &ND_Sim_mu_end_pz,         "ND_Sim_mu_end_pz/D");
-  effTreeFD->Branch("ND_Sim_mu_end_E",                           &ND_Sim_mu_end_E,          "ND_Sim_mu_end_E/D");
-  effTreeFD->Branch("ND_Sim_hadron_contain_result_before_throw", &ND_Sim_hadron_contain_result_before_throw);          // nested vector
-  effTreeFD->Branch("ND_Sim_hadron_throw_result",                &ND_Sim_hadron_throw_result);
-  effTreeFD->Branch("ND_Sim_hadronic_Edep_a2",                   &ND_Sim_hadronic_Edep_a2,  "ND_Sim_hadronic_Edep_a2/D"); // entries = written evts
-  effTreeFD->Branch("FD_Sim_mu_start_vy",                        &FD_Sim_mu_start_vy,       "FD_Sim_mu_start_vy/D");
-  effTreeFD->Branch("FD_Sim_mu_start_vz",                        &FD_Sim_mu_start_vz,       "FD_Sim_mu_start_vz/D");
+  effTreeFD->Branch("ND_Sim_n_hadronic_Edep_a",                  &FD_Sim_n_hadronic_Edep_a,            "FD_Sim_n_hadronic_Edep_a/I");
+  // 1. FD to ND: after earth curvature rotation
+  effTreeFD->Branch("ND_RandomVtx_Sim_mu_start_v",               ND_RandomVtx_Sim_mu_start_v,       "ND_RandomVtx_Sim_mu_start_v[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_RandomVtx_Sim_mu_end_v",                 ND_RandomVtx_Sim_mu_end_v,         "ND_RandomVtx_Sim_mu_end_v[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_RandomVtx_Sim_mu_start_p",               ND_RandomVtx_Sim_mu_start_p,       "ND_RandomVtx_Sim_mu_start_p[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_Sim_hadronic_Edep_a2",                   &ND_Sim_hadronic_Edep_a2,          "ND_Sim_hadronic_Edep_a2/D"); // entries = written evts
+  effTreeFD->Branch("ND_RandomVtx_Sim_hadronic_hit_xyz",         &ND_RandomVtx_Sim_hadronic_hit);
 
+  // 2. ND: move back to the beam center
+  effTreeFD->Branch("ND_OnAxis_Sim_mu_start_v",               ND_OnAxis_Sim_mu_start_v,       "ND_OnAxis_Sim_mu_start_v[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OnAxis_Sim_mu_end_v",                 ND_OnAxis_Sim_mu_end_v,         "ND_OnAxis_Sim_mu_end_v[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OnAxis_Sim_mu_start_p",               ND_OnAxis_Sim_mu_start_p,       "ND_OnAxis_Sim_mu_start_p[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OnAxis_Sim_hadronic_hit",             &ND_OnAxis_Sim_hadronic_hit);
+  // 3. ND to ND: translate from OnAxis to OffAxis
+  effTreeFD->Branch("ND_off_axis_pos_vec",                       &ND_off_axis_pos_vec);                             // vector<double>: entries = written evts * ND_off_axis_pos_steps
+  effTreeFD->Branch("ND_vtx_vx_vec",                             &ND_vtx_vx_vec);
+  effTreeFD->Branch("ND_OffAxis_Unrotated_Sim_mu_start_v",               ND_OffAxis_Unrotated_Sim_mu_start_v,       "ND_OffAxis_Unrotated_Sim_mu_start_v[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OffAxis_Unrotated_Sim_mu_end_v",                 ND_OffAxis_Unrotated_Sim_mu_end_v,         "ND_OffAxis_Unrotated_Sim_mu_end_v[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OffAxis_Unrotated_Sim_mu_start_p",               ND_OffAxis_Unrotated_Sim_mu_start_p,       "ND_OffAxis_Unrotated_Sim_mu_start_p[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OffAxis_Unrotated_Sim_hadronic_hit_xyz",         &ND_OffAxis_Unrotated_Sim_hadronic_hit);
+  // 4. ND: get after eigen rotated vectors for step 4
+  effTreeFD->Branch("ND_OffAxis_Sim_mu_start_v",               ND_OffAxis_Sim_mu_start_v,       "ND_OffAxis_Sim_mu_start_v[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OffAxis_Sim_mu_end_v",                 ND_OffAxis_Sim_mu_end_v,         "ND_OffAxis_Sim_mu_end_v[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OffAxis_Sim_mu_start_p",               ND_OffAxis_Sim_mu_start_p,       "ND_OffAxis_Sim_mu_start_p[3]/D");   // entries = written evts*3
+  effTreeFD->Branch("ND_OffAxis_Sim_mu_start_E",               &ND_OffAxis_Sim_mu_start_E,       "ND_OffAxis_Sim_mu_start_E/D");
+  effTreeFD->Branch("ND_OffAxis_Sim_hadronic_hit_xyz",         &ND_OffAxis_Sim_hadronic_hit);
+  // 5. ND: generate random throws
+  effTreeFD->Branch("ND_Sim_hadron_throw_result",              &ND_Sim_hadron_throw_result);
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //
   //
   // A separate tree to store translations and rotations of throws
   // which will be applied to leptons before NN training
   //
-
   vector<float> throwVtxY;
   vector<float> throwVtxZ;
   vector<float> throwRot;
@@ -255,23 +299,29 @@ int main(){
 
   // Mean neutrino production point (beam coordinate) on z axis as a function of ND off-axis position
   TGraph* gDecayZ = new TGraph(14, OffAxisPoints, meanPDPZ);
-
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
   //
   // Get beam parameters: eventually need to read from XML file: which XML for FD? same as ND?
   //
-
   // Clockwise rotate beamline around ND local x axis
   double beamLineRotation = -0.101;           // unit: rad
+
+  // Put back into beam center(0.0, 0.05387, 6.66) tanslation 1 [cm]
+  ND_OnAxis_Sim_mu_start_v[0] = 0.0*100.;
+  ND_OnAxis_Sim_mu_start_v[1] = 0.05387*100.;
+  ND_OnAxis_Sim_mu_start_v[2] = 6.6*100.;
   // Coordinate transformation, units: meters
-  double beamRefDetCoord[3] = {0., 0., 15.5}; // (0, 0, 0) is ND detector origin
+  double beamRefDetCoord[3] = {ND_OnAxis_Sim_mu_start_v[0]/100., ND_OnAxis_Sim_mu_start_v[1]/100., ND_OnAxis_Sim_mu_start_v[2]/100.}; // (0, 0, 0) is ND detector origin
   double detRefBeamCoord[3] = {0., 0., 574.}; // (0, 0, 0) is beam origin
+
   // Calculate neutrino production x in detector coordinate, y/z later as they depend on ND off-axis position
   decayXdetCoord = beamRefDetCoord[0] - detRefBeamCoord[0];
-
   //
   // Initialize geometric efficiency module
   //
-
   geoEff * eff = new geoEff(314, false); // set verbose to true for debug
   eff->setNthrows(N_throws);
   // Rotate w.r.t. neutrino direction, rather than fixed beam direction
@@ -298,34 +348,45 @@ int main(){
   eff->setOffsetY(offset[1]);
   eff->setOffsetZ(offset[2]);
 
-  //Add hist of veto E
+  // Set On-Axis vertex where center beam cross
+  eff->setOnAxisVertex(ND_OnAxis_Sim_mu_start_v[0],ND_OnAxis_Sim_mu_start_v[1],ND_OnAxis_Sim_mu_start_v[2]);
+
+  // Add hist of veto E
   TH1F *hist_vetoEnergyFD = new TH1F("hist_vetoEnergyFD", "hist_vetoEnergyFD", 1500, 0, 1500);
-  TH1F *hist_vetoEnergyFD_new = new TH1F("hist_vetoEnergyFD_new", "hist_vetoEnergyFD_new", 1500, 0, 1500);
-
-
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //
+  // Add output txt file
+  ofstream myfile;
+   myfile.open ("Output_FDGeoEff_DataCheck.txt");
+  //
   //
   // Loop over FD events
   //
-
   nentries = t->GetEntries();
   std::cout << "Tot evts: " << nentries << std::endl;
-  for ( int ientry = 0; ientry < nentries; ientry++ ) {
-
+  myfile << "Tot evts: " << nentries << "\n";
+  for ( int ientry = 0; ientry < nentries; ientry++ )
+  {
     t->GetEntry(ientry);
-    if ( ientry%10000 == 0 ) std::cout << "Looking at entry " << ientry << ", FD_run: " << FD_Run << ", FD_subrun: " << FD_SubRun << ", FD_event: " << FD_Event << std::endl;
-
+    if ( ientry%10000 == 0 )
+    {
+      std::cout << "Looking at entry " << ientry << ", FD_run: " << FD_Run << ", FD_subrun: " << FD_SubRun << ", FD_event: " << FD_Event << std::endl;
+      myfile << "Looking at entry " << ientry << ", FD_run: " << FD_Run << ", FD_subrun: " << FD_SubRun << ", FD_event: " << FD_Event << "\n\n\n";
+    }
+    myfile << "\n ientry: " << ientry <<"\n\n";
     //
     // Skip events without muon/hadronic deposits
     //
-
     if ( FD_Sim_nMu == 0 || FD_Sim_n_hadronic_Edep_a == 0 ) continue;
+    myfile << "FD_Sim_n_hadronic_Edep_a: " << FD_Sim_n_hadronic_Edep_a <<"\n";
     if ( FD_CCNC_truth == 1) continue;   // only use CC events
     if ( abs(FD_neuPDG) != 14 ) continue;       // only use muon neu
-
     //
     // Calculate total hadron E in FD veto region
     //
-    // Old veto region restriction
     vetoEnergyFD = 0.;
     // Loop over hadron E deposits
     for ( int ihadronhit = 0; ihadronhit < FD_Sim_n_hadronic_Edep_a; ihadronhit++ ){
@@ -346,44 +407,17 @@ int main(){
     //add a vetoEnergyFD histogram in the root file
 
     hist_vetoEnergyFD->Fill(vetoEnergyFD);
-
-
-
-    // New veto region restriction
-    vetoEnergyFD_new = 0.;
-    // Loop over hadron E deposits
-    for ( int ihadronhit = 0; ihadronhit < FD_Sim_n_hadronic_Edep_a; ihadronhit++ ){
-
-      // Veto region size: 30 cm from the active volume
-      if ( ( FD_Sim_hadronic_hit_x_a->at(ihadronhit) < FDActiveVol_min[0] + 30 ) ||
-           ( FD_Sim_hadronic_hit_y_a->at(ihadronhit) < FDActiveVol_min[1] + 30 ) ||
-           ( FD_Sim_hadronic_hit_z_a->at(ihadronhit) < FDActiveVol_min[2] + 30 ) ||
-           ( FD_Sim_hadronic_hit_x_a->at(ihadronhit) > FDActiveVol_max[0] - 30 ) ||
-           ( FD_Sim_hadronic_hit_y_a->at(ihadronhit) > FDActiveVol_max[1] - 30 ) ||
-           ( FD_Sim_hadronic_hit_z_a->at(ihadronhit) > FDActiveVol_max[2] - 30 )
-         ){
-           vetoEnergyFD_new += FD_Sim_hadronic_hit_Edep_a2->at(ihadronhit);
-      } // end if hadron deposit in FD veto region
-
-    } // end loop over hadron E deposits
-    ND_vetoEnergyFD_new = vetoEnergyFD_new;
-    //add a vetoEnergyFD histogram in the root file
-    hist_vetoEnergyFD_new->Fill(vetoEnergyFD_new);
-
-    /*// Add vetoEnergyFD ntuple before threshold
-    ND_vetoEnergyFD = vetoEnergyFD;
-    vetoE->Fill();*/
-
     //
     // Skip FD event if the total hadron E in veto region exceeds vetoEnergy [MeV]
     //
     if ( vetoEnergyFD > 30 ) continue; // 30 MeV
-
+    //
     // Renew throws every 100th written event to save file size, i.e., if N = 128,
     // for written evt 0-99:    same 128 transformations for each event,
     // for written evt 100-199: same but renewed 128 transformations for each evt
     // so on so forth...
     // These transformations will be applied to leptons in the event, so need to keep track of iwritten
+    //
     if ( iwritten % 100 == 0 ) {
 
       // Produce N throws defined at setNthrows(N)
@@ -397,209 +431,355 @@ int main(){
       throwRot  = eff->getCurrentThrowRotations();
       ThrowsFD->Fill();
     }
-
     //
-    // Use relative coordinate for FD event (relative to muon start position)
-    // Set muon start pos in ND to (random x/stepwise increased x, random y, random z),
-    // eventually should set this for actual event vtx, not mu start pos
+    //------------------------------------------------------------------------------
+    //------------------------------------------------------------------------------
+    //------------------------------------------------------------------------------
     //
-
-
-    /*
-    // Use random y/z for each FD evt in ND volume, it will be randomly translated later anyway
-    TRandom3 *r3_mu_start_vy_nd = new TRandom3(); // Initialize random number generator, put inside the event loop so each event is different
-    r3_mu_start_vy_nd->SetSeed(0);                // Set the seed (required to avoid repeated random numbers in each sequence)
-    ND_Sim_mu_start_vy = r3_mu_start_vy_nd->Uniform( NDActiveVol_min[1], NDActiveVol_max[1] );
-
-    TRandom3 *r3_mu_start_vz_nd = new TRandom3();
-    r3_mu_start_vz_nd->SetSeed(0);
-    ND_Sim_mu_start_vz = r3_mu_start_vz_nd->Uniform( NDActiveVol_min[2], NDActiveVol_max[2] );
-
-    ND_Sim_mu_end_vy   = FD_Sim_mu_end_vy - FD_Sim_mu_start_vy + ND_Sim_mu_start_vy; // w.r.t. mu start random y in ND
-    ND_Sim_mu_end_vz   = FD_Sim_mu_end_vz - FD_Sim_mu_start_vz + ND_Sim_mu_start_vz;
-    */
-
-
-    // Local y-z axes in FD and ND are rotated due to Earth curvature, x direction is not change
-    // FD event coordinates, if unchanged, would represent different event in ND coordinate sys.
-    // Apply an active transformation matrix R_x(theta): rotate each point counterclockwise by theta around x-axis in ND
-    // theta is 2*|beamLineRotation|
-    // Transform FD relative coordinate, x coordinate unchanged
-    //
-    //              [ 1          0             0
-    // R_x(theta) =   0      cos(theta)   -sin(theta)
-    //                0      sin(theta)    cos(theta) ]
-    //
-
+    // 1. FD to ND: after earth curvature rotation
     // Rotation affects mu start/end position and hadron positions below
-    ND_Sim_mu_start_vy = cos( 2*abs(beamLineRotation) )*FD_Sim_mu_start_vy - sin( 2*abs(beamLineRotation) )*FD_Sim_mu_start_vz;
-    ND_Sim_mu_start_vz = sin( 2*abs(beamLineRotation) )*FD_Sim_mu_start_vy + cos( 2*abs(beamLineRotation) )*FD_Sim_mu_start_vz;
+    // Mu_start_v
+    double FD_Sim_mu_start_v[3] = {FD_Sim_mu_start_vx, FD_Sim_mu_start_vy, FD_Sim_mu_start_vz};
+    for(int i=0; i<3; i++)
+    {
+      ND_RandomVtx_Sim_mu_start_v[i] = eff->getEarthCurvature(FD_Sim_mu_start_v, beamLineRotation, i);
+    }
+    myfile << "FD_Sim_mu_start_vx[cm]: " << FD_Sim_mu_start_vx << "\n";
+    myfile << "FD_Sim_mu_start_vy[cm]: " << FD_Sim_mu_start_vy << "\n";
+    myfile << "FD_Sim_mu_start_vz[cm]: " << FD_Sim_mu_start_vz << "\n\n";
+    myfile << "ND_RandomVtx_Sim_mu_start_vx[cm]: " << ND_RandomVtx_Sim_mu_start_v[0] << "\n";
+    myfile << "ND_RandomVtx_Sim_mu_start_vy[cm]: " << ND_RandomVtx_Sim_mu_start_v[1] << "\n";
+    myfile << "ND_RandomVtx_Sim_mu_start_vz[cm]: " << ND_RandomVtx_Sim_mu_start_v[2] << "\n\n";
+    // Mu_end_v
+    double FD_Sim_mu_end_v[3] = {FD_Sim_mu_end_vx, FD_Sim_mu_end_vy, FD_Sim_mu_end_vz};
+    for(int i=0; i<3; i++)
+    {
+      ND_RandomVtx_Sim_mu_end_v[i] = eff->getEarthCurvature(FD_Sim_mu_end_v, beamLineRotation, i);
+    }
+    myfile << "FD_Sim_mu_end_vx[cm]: " << FD_Sim_mu_end_vx << "\n";
+    myfile << "FD_Sim_mu_end_vy[cm]: " << FD_Sim_mu_end_vy << "\n";
+    myfile << "FD_Sim_mu_end_vz[cm]: " << FD_Sim_mu_end_vz << "\n";
+    double FD_Sim_mu_v_end_start = eff->getDistance(FD_Sim_mu_start_v,FD_Sim_mu_end_v);
+    myfile << "Distance between FD_Sim_mu_v_end_start[cm]: " << FD_Sim_mu_v_end_start << "\n\n";
+    myfile << "ND_RandomVtx_Sim_mu_end_vx[cm]: " << ND_RandomVtx_Sim_mu_end_v[0] << "\n";
+    myfile << "ND_RandomVtx_Sim_mu_end_vy[cm]: " << ND_RandomVtx_Sim_mu_end_v[1] << "\n";
+    myfile << "ND_RandomVtx_Sim_mu_end_vz[cm]: " << ND_RandomVtx_Sim_mu_end_v[2] << "\n";
+    double ND_RandomVtx_Sim_mu_v_end_start = eff->getDistance(ND_RandomVtx_Sim_mu_start_v,ND_RandomVtx_Sim_mu_end_v);
+    myfile << "Distance between ND_RandomVtx_Sim_mu_v_end_start[cm]: " << ND_RandomVtx_Sim_mu_v_end_start << "\n\n";
 
-    ND_Sim_mu_end_vy = cos( 2*abs(beamLineRotation) )*FD_Sim_mu_end_vy - sin( 2*abs(beamLineRotation) )*FD_Sim_mu_end_vz;
-    ND_Sim_mu_end_vz = sin( 2*abs(beamLineRotation) )*FD_Sim_mu_end_vy + cos( 2*abs(beamLineRotation) )*FD_Sim_mu_end_vz;
+    // Mu_start_p
+    double FD_Sim_mu_start_p[3] = {FD_Sim_mu_start_px, FD_Sim_mu_start_py, FD_Sim_mu_start_pz};
+    for(int i=0; i<3; i++)
+    {
+      ND_RandomVtx_Sim_mu_start_p[i] = eff->getEarthCurvature(FD_Sim_mu_start_p, beamLineRotation, i);
+    }
+    ND_RandomVtx_Sim_mu_start_p_total = eff->getTotalMomentum(ND_RandomVtx_Sim_mu_start_p);
+    double FD_Sim_mu_start_p_total = eff->getTotalMomentum(FD_Sim_mu_start_p);
+    myfile << "FD_Sim_mu_start_px[cm]: " << FD_Sim_mu_start_px << "\n";
+    myfile << "FD_Sim_mu_start_py[cm]: " << FD_Sim_mu_start_py << "\n";
+    myfile << "FD_Sim_mu_start_pz[cm]: " << FD_Sim_mu_start_pz << "\n";
+    myfile << "FD_Sim_mu_start_p_total[GeV]: " << FD_Sim_mu_start_p_total << "\n\n";
+    myfile << "ND_RandomVtx_Sim_mu_start_px[GeV]: " << ND_RandomVtx_Sim_mu_start_p[0] << "\n";
+    myfile << "ND_RandomVtx_Sim_mu_start_py[GeV]: " << ND_RandomVtx_Sim_mu_start_p[1] << "\n";
+    myfile << "ND_RandomVtx_Sim_mu_start_pz[GeV]: " << ND_RandomVtx_Sim_mu_start_p[2] << "\n";
+    myfile << "ND_RandomVtx_Sim_mu_start_p_total[GeV]: " << ND_RandomVtx_Sim_mu_start_p_total << "\n\n";
 
     // Initialize for the event
-    ND_Sim_mu_start_vx.clear();
-    ND_Sim_mu_end_vx.clear();
-    ND_Sim_hadron_contain_result_before_throw.clear();
     ND_Sim_hadron_throw_result.clear();
-
     // Branches that are not affected by ND off axis position and vtx x (loops below)
     ND_Gen_numu_E      = FD_Gen_numu_E;
-
-    // X momentum is not affected by coordinate rotation
-    ND_Sim_mu_start_px = FD_Sim_mu_start_px;
-    ND_Sim_mu_end_px   = FD_Sim_mu_end_px;
-
-    // Rotation affects mu start/end momentum vector in Y and Z axes, using the same rotation matrix below.
-    ND_Sim_mu_start_py = cos( 2*abs(beamLineRotation) )*FD_Sim_mu_start_py - sin( 2*abs(beamLineRotation) )*FD_Sim_mu_start_pz;
-    ND_Sim_mu_start_pz = sin( 2*abs(beamLineRotation) )*FD_Sim_mu_start_py + cos( 2*abs(beamLineRotation) )*FD_Sim_mu_start_pz;
-    ND_Sim_mu_end_py   = cos( 2*abs(beamLineRotation) )*FD_Sim_mu_end_py - sin( 2*abs(beamLineRotation) )*FD_Sim_mu_end_pz;
-    ND_Sim_mu_end_pz   = sin( 2*abs(beamLineRotation) )*FD_Sim_mu_end_py + cos( 2*abs(beamLineRotation) )*FD_Sim_mu_end_pz;
-
-    // Energy is a scalar, doesn't change
-    ND_Sim_mu_start_E  = FD_Sim_mu_start_E;
-    ND_Sim_mu_end_E    = FD_Sim_mu_end_E;
     ND_Sim_hadronic_Edep_a2 = FD_Sim_hadronic_Edep_a2;
 
-    //
-    // Two options for setting ND off-axis position
-    //
-    // Option 1: randomize ND off-axis position once for each event
-    // Option 2: stepwise increments along ND detector hall off axis range
-    //
-    // If only want option 1, set random_ND_off_axis_pos to true in Helpers.h; default is false (use both options)
-    //
 
-    // Initialize random number generator
-    // This needs to be inside the event loop to make sure each event has a different random number
-    TRandom3 *r3_OffAxisPoint = new TRandom3();
-    // Set the seed (required to avoid repeated random numbers in each sequence)
-    r3_OffAxisPoint->SetSeed(0);
-    ND_off_axis_pos_vec.at(0) = r3_OffAxisPoint->Uniform(OffAxisPoints[0], OffAxisPoints[13]);
-
-    if (verbose) std::cout << "random OffAxisPoint [meters]: " << ND_off_axis_pos_vec.at(0) << std::endl;
-
+    for ( int ihadronhit = 0; ihadronhit < FD_Sim_n_hadronic_Edep_a; ihadronhit++ ){
+      double FD_Sim_hadronic_hit[3] = {FD_Sim_hadronic_hit_x_a->at(ihadronhit),FD_Sim_hadronic_hit_y_a->at(ihadronhit), FD_Sim_hadronic_hit_z_a->at(ihadronhit)};
+      for (int i =0; i<3;i++)
+      {
+        ND_RandomVtx_Sim_hadronic_hit_xyz.emplace_back((float)eff->getEarthCurvature(FD_Sim_hadronic_hit, beamLineRotation, i));
+      }
+      ND_RandomVtx_Sim_hadronic_hit.emplace_back(ND_RandomVtx_Sim_hadronic_hit_xyz);
+      ND_RandomVtx_Sim_hadronic_hit_xyz.clear();
+    }
     //
-    // Similarly, two options for setting event vtx x position
+    //------------------------------------------------------------------------------
+    //------------------------------------------------------------------------------
+    //------------------------------------------------------------------------------
     //
-    // Option 1: randomize x once for each event
-    // Option 2: stepwise increments along ND local width in x: -2m to 2m
+    // 2. ND: move back to the beam center
+    // ND_OnAxis_Sim_mu_end_v
+    // double geoEff::getTranslations(double v_bf[3], double vtx_bf[3], double vtx_af[3], int dim)
+    for(int i=0; i<3; i++)
+    {
+      ND_OnAxis_Sim_mu_end_v[i] = eff->getTranslations(ND_RandomVtx_Sim_mu_end_v, ND_RandomVtx_Sim_mu_start_v, ND_OnAxis_Sim_mu_start_v, i);
+    }
+    myfile << "ND_OnAxis_Sim_mu_start_vx[cm]: " << ND_OnAxis_Sim_mu_start_v[0] << "\n";
+    myfile << "ND_OnAxis_Sim_mu_start_vy[cm]: " << ND_OnAxis_Sim_mu_start_v[1] << "\n";
+    myfile << "ND_OnAxis_Sim_mu_start_vz[cm]: " << ND_OnAxis_Sim_mu_start_v[2] << "\n\n";
+    myfile << "ND_OnAxis_Sim_mu_end_vx[cm]: " << ND_OnAxis_Sim_mu_end_v[0] << "\n";
+    myfile << "ND_OnAxis_Sim_mu_end_vy[cm]: " << ND_OnAxis_Sim_mu_end_v[1] << "\n";
+    myfile << "ND_OnAxis_Sim_mu_end_vz[cm]: " << ND_OnAxis_Sim_mu_end_v[2] << "\n";
+    double ND_OnAxis_Sim_mu_v_end_start = eff->getDistance(ND_OnAxis_Sim_mu_start_v,ND_OnAxis_Sim_mu_end_v);
+    myfile << "Distance between ND_OnAxis_Sim_mu_v_end_start[cm]: " << ND_OnAxis_Sim_mu_v_end_start << "\n\n";
+    // ND_OnAxis_Sim_mu_start_p
+    for(int i=0; i<3; i++)
+    {
+      ND_OnAxis_Sim_mu_start_p[i] = eff->RemainUnchanged(ND_RandomVtx_Sim_mu_start_p[i]);
+    }
+    ND_OnAxis_Sim_mu_start_p_total = eff->getTotalMomentum(ND_OnAxis_Sim_mu_start_p);
+    myfile << "ND_OnAxis_Sim_mu_start_px[cm]: " << ND_OnAxis_Sim_mu_start_p[0] << "\n";
+    myfile << "ND_OnAxis_Sim_mu_start_py[cm]: " << ND_OnAxis_Sim_mu_start_p[1] << "\n";
+    myfile << "ND_OnAxis_Sim_mu_start_pz[cm]: " << ND_OnAxis_Sim_mu_start_p[2] << "\n";
+    myfile << "ND_OnAxis_Sim_mu_start_p_total[GeV]: " << ND_OnAxis_Sim_mu_start_p_total << "\n\n";
+    // ND_OnAxis_Sim_hadronic_hit
+    for ( int ihadronhit = 0; ihadronhit < FD_Sim_n_hadronic_Edep_a; ihadronhit++ ){
+      double ND_RandomVtx_Sim_hadronic_hit_array[3] = {ND_RandomVtx_Sim_hadronic_hit[ihadronhit][0],ND_RandomVtx_Sim_hadronic_hit[ihadronhit][1],ND_RandomVtx_Sim_hadronic_hit[ihadronhit][2]};
+      for (int i =0; i<3;i++)
+      {
+        ND_OnAxis_Sim_hadronic_hit_xyz.emplace_back((float)eff->getTranslations(ND_RandomVtx_Sim_hadronic_hit_array, ND_RandomVtx_Sim_mu_start_v, ND_OnAxis_Sim_mu_start_v, i));
+      }
+      ND_OnAxis_Sim_hadronic_hit.emplace_back(ND_OnAxis_Sim_hadronic_hit_xyz);
+      ND_OnAxis_Sim_hadronic_hit_xyz.clear();
+    }
     //
-    // If only want option 1, set random_vtx_vx to true in Helpers.h; default is false (use both options)
-    //
-
-    TRandom3 *r3_vtx_x = new TRandom3();
-    r3_vtx_x->SetSeed(0);
-    ND_vtx_vx_vec.at(0) = r3_vtx_x->Uniform(ND_local_x_min, ND_local_x_max);
-
-    if (verbose) std::cout << "random vtx_x [cm]: " << ND_vtx_vx_vec.at(0) << std::endl;
-
+    //------------------------------------------------------------------------------
+    //------------------------------------------------------------------------------
+    //------------------------------------------------------------------------------
     //
     // Loop over ND_off_axis_pos_vec: random off_axis_pos or every ND_off_axis_pos_stepsize
     // Don't put it outside event loop to avoid looping over all events multiple times
     //
 
     int ND_off_axis_pos_counter = 0;
-    for ( double i_ND_off_axis_pos : ND_off_axis_pos_vec ) {
-
-      // Skip the stepwise increased option if only want a random off axis ND position per event to save file size
-      if ( random_ND_off_axis_pos && ND_off_axis_pos_counter != 0 ) continue;
-
+    for ( double i_ND_off_axis_pos : ND_off_axis_pos_vec )
+    {
       ND_off_axis_pos_counter++;
-
-      // Interpolate event neutrino production point (beam coordinate)
-      decayZbeamCoord = gDecayZ->Eval( i_ND_off_axis_pos - detRefBeamCoord[0] );
-
-      // Calculate neutrino production point in detector coordinate
-      decayYdetCoord = beamRefDetCoord[1] - detRefBeamCoord[1]*cos(beamLineRotation) + ( decayZbeamCoord - detRefBeamCoord[2] )*sin(beamLineRotation);
-      decayZdetCoord = beamRefDetCoord[2] + detRefBeamCoord[1]*sin(beamLineRotation) + ( decayZbeamCoord - detRefBeamCoord[2] )*cos(beamLineRotation);
-      // Set production point in unit: cm
-      eff->setDecayPos(decayXdetCoord*100., decayYdetCoord*100., decayZdetCoord*100.);
-
-      if (verbose) std::cout << "nd off_axis x #" << ND_off_axis_pos_counter << ": " << i_ND_off_axis_pos << " m" << std::endl;
-
       //
       // Loop over vtx x: random x or stepwise increased x
       // Don't put it outside event loop to avoid looping over all events multiple times
       //
-
       int vtx_vx_counter = 0;
       hadron_throw_result_vec_for_vtx_vx.clear(); // Need to initialize before loop over vtx x vec
-      hadron_contain_result_before_throw_vec_for_vtx_vx.clear();
-
-      for ( double i_vtx_vx : ND_vtx_vx_vec ) {
-
-        // Skip the stepwise increased option if only want a random evt vtx x to save file size
-        if ( random_vtx_vx && vtx_vx_counter != 0 ) continue;
+      //
+      //------------------------------------------------------------------------------
+      //------------------------------------------------------------------------------
+      //------------------------------------------------------------------------------
+      //
+      for ( double i_vtx_vx : ND_vtx_vx_vec )
+      {
 
         vtx_vx_counter++;
 
-        // ND off-axis position does not affect evt vx, so only fill branches below once when loop over ND off-axis vec
-        if ( ND_off_axis_pos_counter == 1 ) {
-          ND_Sim_mu_start_vx.emplace_back( i_vtx_vx );
-          ND_Sim_mu_end_vx.emplace_back( FD_Sim_mu_end_vx - FD_Sim_mu_start_vx + i_vtx_vx ); // w.r.t. mu start x
+        // Interpolate event neutrino production point (beam coordinate)
+        decayZbeamCoord = gDecayZ->Eval( i_ND_off_axis_pos + i_vtx_vx - detRefBeamCoord[0] );
+
+        // Calculate neutrino production point in detector coordinate
+        decayYdetCoord = beamRefDetCoord[1] - detRefBeamCoord[1]*cos(beamLineRotation) + ( decayZbeamCoord - detRefBeamCoord[2] )*sin(beamLineRotation);
+        decayZdetCoord = beamRefDetCoord[2] + detRefBeamCoord[1]*sin(beamLineRotation) + ( decayZbeamCoord - detRefBeamCoord[2] )*cos(beamLineRotation);
+        // Set production point in unit: cm
+        eff->setDecayPos(decayXdetCoord*100., decayYdetCoord*100., decayZdetCoord*100.);
+
+        if (verbose) std::cout << "nd off_axis x #" << ND_off_axis_pos_counter << ": " << i_ND_off_axis_pos << " m" << std::endl;
+        myfile << "Nd Off-Axis x #" << ND_off_axis_pos_counter << "[cm]: " << i_ND_off_axis_pos << "\n";
+        myfile << "Nd Off-Axis x in detector #" << vtx_vx_counter << "[cm]: " << i_vtx_vx << "\n";
+        myfile << "Event position #" << vtx_vx_counter << "[cm]: " << i_ND_off_axis_pos + i_vtx_vx << "\n\n";
+
+        //
+        //------------------------------------------------------------------------------
+        //------------------------------------------------------------------------------
+        //------------------------------------------------------------------------------
+        //
+        // 3. ND to ND: translate from OnAxis to OffAxis
+        //
+        // ND_OffAxis_Unrotated_Sim_mu_start_v
+        ND_OffAxis_Unrotated_Sim_mu_start_v[0] = ND_OnAxis_Sim_mu_start_v[0] + i_ND_off_axis_pos + i_vtx_vx;
+        ND_OffAxis_Unrotated_Sim_mu_start_v[1] = eff->RemainUnchanged(ND_OnAxis_Sim_mu_start_v[1]);
+        ND_OffAxis_Unrotated_Sim_mu_start_v[2] = eff->RemainUnchanged(ND_OnAxis_Sim_mu_start_v[2]);
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_start_vx[cm]: " << ND_OffAxis_Unrotated_Sim_mu_start_v[0] << "\n";
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_start_vy[cm]: " << ND_OffAxis_Unrotated_Sim_mu_start_v[1] << "\n";
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_start_vz[cm]: " << ND_OffAxis_Unrotated_Sim_mu_start_v[2] << "\n\n";
+        // ND_OffAxis_Unrotated_Sim_mu_end_v
+        // double geoEff::getTranslations(double v_bf[3], double vtx_bf[3], double vtx_af[3], int dim)
+        for(int i=0; i<3; i++)
+        {
+          ND_OffAxis_Unrotated_Sim_mu_end_v[i] = eff->getTranslations(ND_OnAxis_Sim_mu_end_v, ND_OnAxis_Sim_mu_start_v, ND_OffAxis_Unrotated_Sim_mu_start_v, i);
+        }
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_end_vx[cm]: " << ND_OffAxis_Unrotated_Sim_mu_end_v[0] << "\n";
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_end_vy[cm]: " << ND_OffAxis_Unrotated_Sim_mu_end_v[1] << "\n";
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_end_vz[cm]: " << ND_OffAxis_Unrotated_Sim_mu_end_v[2] << "\n";
+        double ND_OffAxis_Unrotated_Sim_mu_v_end_start = eff->getDistance(ND_OffAxis_Unrotated_Sim_mu_end_v,ND_OffAxis_Unrotated_Sim_mu_start_v);
+        myfile << "Distance between ND_OffAxis_Unrotated_Sim_mu_v_end_start[cm]: " << ND_OffAxis_Unrotated_Sim_mu_v_end_start << "\n\n";
+        // ND_OffAxis_Unrotated_Sim_mu_start_p
+        for(int i=0; i<3; i++)
+        {
+          ND_OffAxis_Unrotated_Sim_mu_start_p[i] = eff->RemainUnchanged(ND_OnAxis_Sim_mu_start_p[i]);
+        }
+        ND_OffAxis_Unrotated_Sim_mu_start_p_total = eff->getTotalMomentum(ND_OffAxis_Unrotated_Sim_mu_start_p);
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_start_px[cm]: " << ND_OffAxis_Unrotated_Sim_mu_start_p[0] << "\n";
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_start_py[cm]: " << ND_OffAxis_Unrotated_Sim_mu_start_p[1] << "\n";
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_start_pz[cm]: " << ND_OffAxis_Unrotated_Sim_mu_start_p[2] << "\n";
+        myfile << "ND_OffAxis_Unrotated_Sim_mu_start_p_total[GeV]: " << ND_OffAxis_Unrotated_Sim_mu_start_p_total << "\n\n";
+        // ND_OffAxis_Unrotated_Sim_hadronic_hit
+        for ( int ihadronhit = 0; ihadronhit < FD_Sim_n_hadronic_Edep_a; ihadronhit++ ){
+          double ND_OnAxis_Sim_hadronic_hit_array[3] = {ND_OnAxis_Sim_hadronic_hit[ihadronhit][0],ND_OnAxis_Sim_hadronic_hit[ihadronhit][1],ND_OnAxis_Sim_hadronic_hit[ihadronhit][2]};
+          for (int i =0; i<3;i++)
+          {
+            ND_OffAxis_Unrotated_Sim_hadronic_hit_xyz.emplace_back((float)eff->getTranslations(ND_OnAxis_Sim_hadronic_hit_array, ND_OnAxis_Sim_mu_start_v, ND_OffAxis_Unrotated_Sim_mu_start_v, i));
+          }
+          ND_OffAxis_Unrotated_Sim_hadronic_hit.emplace_back(ND_OffAxis_Unrotated_Sim_hadronic_hit_xyz);
+          ND_OffAxis_Unrotated_Sim_hadronic_hit_xyz.clear();
+        }
+        //
+        //------------------------------------------------------------------------------
+        //------------------------------------------------------------------------------
+        //------------------------------------------------------------------------------
+        //
+        // 4. ND: get after eigen rotated vectors for step 3
+        // ND_OffAxis_Sim_mu_start_v
+        ND_OffAxis_Sim_mu_start_v[0] = eff->RemainUnchanged(ND_OffAxis_Unrotated_Sim_mu_start_v[0]);
+        ND_OffAxis_Sim_mu_start_v[1] = eff->RemainUnchanged(ND_OffAxis_Unrotated_Sim_mu_start_v[1]);
+        ND_OffAxis_Sim_mu_start_v[2] = eff->RemainUnchanged(ND_OffAxis_Unrotated_Sim_mu_start_v[2]);
+        myfile << "ND_OffAxis_Sim_mu_start_v[cm]: " << ND_OffAxis_Sim_mu_start_v[0] << "\n";
+        myfile << "ND_OffAxis_Sim_mu_start_v[cm]: " << ND_OffAxis_Sim_mu_start_v[1] << "\n";
+        myfile << "ND_OffAxis_Sim_mu_start_v[cm]: " << ND_OffAxis_Sim_mu_start_v[2] << "\n\n";
+        // setVertex
+        eff->setOffAxisVertex(ND_OffAxis_Sim_mu_start_v[0], ND_OffAxis_Sim_mu_start_v[1], ND_OffAxis_Sim_mu_start_v[2]);
+        // ND_OffAxis_Sim_mu_end_v
+        eff->setMuEndV(ND_OffAxis_Unrotated_Sim_mu_end_v[0],ND_OffAxis_Unrotated_Sim_mu_end_v[1],ND_OffAxis_Unrotated_Sim_mu_end_v[2]);
+        for(int i=0; i<3; i++)
+        {
+          ND_OffAxis_Sim_mu_end_v[i] = eff->getOffAxisMuEndV(i);
+        }
+        myfile << "ND_OffAxis_Sim_mu_end_v[cm]: " << ND_OffAxis_Sim_mu_end_v[0] << "\n";
+        myfile << "ND_OffAxis_Sim_mu_end_v[cm]: " << ND_OffAxis_Sim_mu_end_v[1] << "\n";
+        myfile << "ND_OffAxis_Sim_mu_end_v[cm]: " << ND_OffAxis_Sim_mu_end_v[2] << "\n";
+        double ND_OffAxis_Sim_mu_v_end_start = eff->getDistance(ND_OffAxis_Sim_mu_end_v,ND_OffAxis_Sim_mu_start_v);
+        myfile << "Distance between ND_OffAxis_Sim_mu_v_end_start[cm]: " << ND_OffAxis_Sim_mu_v_end_start << "\n\n";
+        // ND_OffAxis_Sim_mu_start_p
+        eff->setMuStartP(ND_OffAxis_Unrotated_Sim_mu_start_p[0],ND_OffAxis_Unrotated_Sim_mu_start_p[1],ND_OffAxis_Unrotated_Sim_mu_start_p[2]);
+        for(int i=0; i<3; i++)
+        {
+          ND_OffAxis_Sim_mu_start_p[i] = eff->getOffAxisMuStartP(i);
+        }
+        ND_OffAxis_Sim_mu_start_p_total = eff->getTotalMomentum(ND_OffAxis_Sim_mu_start_p);
+        myfile << "ND_OffAxis_Sim_mu_start_px[cm]: " << ND_OffAxis_Sim_mu_start_p[0] << "\n";
+        myfile << "ND_OffAxis_Sim_mu_start_py[cm]: " << ND_OffAxis_Sim_mu_start_p[1] << "\n";
+        myfile << "ND_OffAxis_Sim_mu_start_pz[cm]: " << ND_OffAxis_Sim_mu_start_p[2] << "\n";
+        myfile << "ND_OffAxis_Sim_mu_start_p_total[GeV]: " << ND_OffAxis_Sim_mu_start_p_total << "\n\n";
+        // ND_OffAxis_Sim_mu_start_E
+        ND_OffAxis_Sim_mu_start_E = FD_Sim_mu_start_E; // Energy of leading mu at start point [GeV]
+        // ND_OffAxis_Sim_hadronic_hit
+        for ( int ihadronhit = 0; ihadronhit < FD_Sim_n_hadronic_Edep_a; ihadronhit++ ){
+          eff->setHadronHitV(ND_OffAxis_Unrotated_Sim_hadronic_hit[ihadronhit][0],ND_OffAxis_Unrotated_Sim_hadronic_hit[ihadronhit][1],ND_OffAxis_Unrotated_Sim_hadronic_hit[ihadronhit][2]);
+          for (int i =0; i<3;i++)
+          {
+            ND_OffAxis_Sim_hadronic_hit_xyz.emplace_back(eff->getOffAxisHadronHitV(i));
+          }
+          ND_OffAxis_Sim_hadronic_hit.emplace_back(ND_OffAxis_Sim_hadronic_hit_xyz);
+          ND_OffAxis_Sim_hadronic_hit_xyz.clear();
         }
 
-        // Evt vtx pos in unit: cm
-        eff->setVertex( i_vtx_vx, ND_Sim_mu_start_vy, ND_Sim_mu_start_vz );
+        // Add output
+        for (int ihadronhit = FD_Sim_n_hadronic_Edep_a-2; ihadronhit < FD_Sim_n_hadronic_Edep_a; ihadronhit++) {
+          myfile << "Hit #" << ihadronhit <<"/" << FD_Sim_n_hadronic_Edep_a << "\n";
+          myfile<<"FD_Sim_hadronic_hit_x[cm]: "<<FD_Sim_hadronic_hit_x_a->at(ihadronhit)<<"\n";
+          myfile<<"FD_Sim_hadronic_hit_y[cm]: "<<FD_Sim_hadronic_hit_y_a->at(ihadronhit)<<"\n";
+          myfile<<"FD_Sim_hadronic_hit_z[cm]: "<<FD_Sim_hadronic_hit_z_a->at(ihadronhit)<<"\n";
+          double FD_Sim_hadronic_hit[3] = {FD_Sim_hadronic_hit_x_a->at(ihadronhit),FD_Sim_hadronic_hit_y_a->at(ihadronhit), FD_Sim_hadronic_hit_z_a->at(ihadronhit)};
+          double FD_Sim_hadronic_hit_end_start = eff->getDistance(FD_Sim_hadronic_hit,FD_Sim_mu_start_v);
+          myfile << "Distance between FD_Sim_hadronic_hit_end_start[cm]:" << FD_Sim_hadronic_hit_end_start <<"\n\n";
 
+          for (int j = 0; j < 3; j++)
+          {
+            myfile << "ND_RandomVtx_Sim_hadronic_hit_["<<j<<"][cm]: "<< ND_RandomVtx_Sim_hadronic_hit[ihadronhit][j] << "\n";
+          }
+          double ND_RandomVtx_Sim_hadronic_hit_array[3] = {ND_RandomVtx_Sim_hadronic_hit[ihadronhit][0],ND_RandomVtx_Sim_hadronic_hit[ihadronhit][1],ND_RandomVtx_Sim_hadronic_hit[ihadronhit][2]};
+          double ND_RandomVtx_Sim_hadronic_hit_end_start = eff->getDistance(ND_RandomVtx_Sim_hadronic_hit_array,ND_RandomVtx_Sim_mu_start_v);
+          myfile << "Distance between ND_RandomVtx_Sim_hadronic_hit_end_start[cm]:" << ND_RandomVtx_Sim_hadronic_hit_end_start <<"\n\n";
+
+          for (int j = 0; j < 3; j++)
+          {
+            myfile << "ND_OnAxis_Sim_hadronic_hit_["<<j<<"][cm]: "<< ND_OnAxis_Sim_hadronic_hit[ihadronhit][j] << "\n";
+          }
+          double ND_OnAxis_Sim_hadronic_hit_array[3] = {ND_OnAxis_Sim_hadronic_hit[ihadronhit][0],ND_OnAxis_Sim_hadronic_hit[ihadronhit][1],ND_OnAxis_Sim_hadronic_hit[ihadronhit][2]};
+          double ND_OnAxis_Sim_hadronic_hit_end_start = eff->getDistance(ND_OnAxis_Sim_hadronic_hit_array,ND_OnAxis_Sim_mu_start_v);
+          myfile << "Distance between ND_OnAxis_Sim_hadronic_hit_end_start[cm]:" << ND_OnAxis_Sim_hadronic_hit_end_start <<"\n\n";
+
+          for (int j = 0; j < 3; j++)
+          {
+            myfile << "ND_OffAxis_Unrotated_Sim_hadronic_hit_["<<j<<"][cm]: "<< ND_OffAxis_Unrotated_Sim_hadronic_hit[ihadronhit][j] << "\n";
+          }
+          double ND_OffAxis_Unrotated_Sim_hadronic_hit_array[3] = {ND_OffAxis_Unrotated_Sim_hadronic_hit[ihadronhit][0],ND_OffAxis_Unrotated_Sim_hadronic_hit[ihadronhit][1],ND_OffAxis_Unrotated_Sim_hadronic_hit[ihadronhit][2]};
+          double ND_OffAxis_Unrotated_Sim_hadronic_hit_end_start = eff->getDistance(ND_OffAxis_Unrotated_Sim_hadronic_hit_array,ND_OffAxis_Unrotated_Sim_mu_start_v);
+          myfile << "Distance between ND_OffAxis_Unrotated_Sim_hadronic_hit_end_start[cm]:" << ND_OffAxis_Unrotated_Sim_hadronic_hit_end_start <<"\n\n";
+
+          for (int j = 0; j < 3; j++)
+          {
+            myfile << "ND_OffAxis_Sim_hadronic_hit_["<<j<<"][cm]: "<< ND_OffAxis_Sim_hadronic_hit[ihadronhit][j] << "\n";
+          }
+          double ND_OffAxis_Sim_hadronic_hit_array[3] = {ND_OffAxis_Sim_hadronic_hit[ihadronhit][0],ND_OffAxis_Sim_hadronic_hit[ihadronhit][1],ND_OffAxis_Sim_hadronic_hit[ihadronhit][2]};
+          double ND_OffAxis_Sim_hadronic_hit_end_start = eff->getDistance(ND_OffAxis_Sim_hadronic_hit_array,ND_OffAxis_Sim_mu_start_v);
+          myfile << "Distance between ND_OffAxis_Sim_hadronic_hit_end_start[cm]:" << ND_OffAxis_Sim_hadronic_hit_end_start <<"\n\n";
+        }
+        //
+        //------------------------------------------------------------------------------
+        //------------------------------------------------------------------------------
+        //------------------------------------------------------------------------------
+        //
+        // 5. ND: generate random throws
+        // Hardon hit positions
         HadronHitEdeps.clear();
         HadronHitPoss.clear();
         HadronHitEdeps.reserve(FD_Sim_n_hadronic_Edep_a);
         HadronHitPoss.reserve(FD_Sim_n_hadronic_Edep_a*3);
-
-        // Need to loop deposits many times as the hadron containment below need vtx x first
+        // Set HadronHitPoss
         for ( int ihadronhit = 0; ihadronhit < FD_Sim_n_hadronic_Edep_a; ihadronhit++ ){
-
-          // Relative to muon start pos in ND coordinate sys: (i_vtx_vx, ND_Sim_mu_start_vy, ND_Sim_mu_start_vz)
-          HadronHitPoss.emplace_back( FD_Sim_hadronic_hit_x_a->at(ihadronhit) - FD_Sim_mu_start_vx + i_vtx_vx ); // w.r.t. mu start x
-          // Again, need to apply R_x(theta) for hadron y/z, do not affect x
-          HadronHitPoss.emplace_back( cos( 2*abs(beamLineRotation) )*( FD_Sim_hadronic_hit_y_a->at(ihadronhit) - FD_Sim_mu_start_vy + ND_Sim_mu_start_vy ) - sin( 2*abs(beamLineRotation) )*( FD_Sim_hadronic_hit_z_a->at(ihadronhit) - FD_Sim_mu_start_vz + ND_Sim_mu_start_vz ) );
-          HadronHitPoss.emplace_back( sin( 2*abs(beamLineRotation) )*( FD_Sim_hadronic_hit_y_a->at(ihadronhit) - FD_Sim_mu_start_vy + ND_Sim_mu_start_vy ) + cos( 2*abs(beamLineRotation) )*( FD_Sim_hadronic_hit_z_a->at(ihadronhit) - FD_Sim_mu_start_vz + ND_Sim_mu_start_vz ) );
-
+          for (int i =0; i<3;i++)
+          {
+            HadronHitPoss.emplace_back(ND_OffAxis_Sim_hadronic_hit[ihadronhit][i]);
+          }
           HadronHitEdeps.emplace_back( FD_Sim_hadronic_hit_Edep_a2->at(ihadronhit) );
+        }
 
-        } // end for loop
-
+        eff->setVertex(ND_OffAxis_Sim_mu_start_v[0], ND_OffAxis_Sim_mu_start_v[1], ND_OffAxis_Sim_mu_start_v[2]);
         eff->setHitSegEdeps(HadronHitEdeps);
         eff->setHitSegPoss(HadronHitPoss); // this is converted hadrom deposit pos in ND coordinate sys.
-
         //
         // Get hadron containment result after everything is set to ND coordinate sys
-        //
-
-        // Store hadron containment result for the FD evt before doing random throws
-        hadron_contain_result_before_throw = eff->getHadronContainmentOrigin();
-        hadron_contain_result_before_throw_vec_for_vtx_vx.emplace_back(hadron_contain_result_before_throw);
-
         // Do random throws regardless whether FD evt is contained in ND volume by setting a false flag
         hadron_throw_result = eff->getHadronContainmentThrows(false); // Every 64 throw results stored into a 64 bit unsigned int: 0101101...
         hadron_throw_result_vec_for_vtx_vx.emplace_back(hadron_throw_result);
 
-        if (verbose) std::cout << "                           vtx x #" << vtx_vx_counter << ": " << i_vtx_vx << " cm, throw result[0][0][0]: " << hadron_throw_result[0][0][0] << std::endl;
+        if (verbose) std::cout << "vtx x #" << vtx_vx_counter << ": " << i_vtx_vx << " cm, throw result[0][0][0]: " << hadron_throw_result[0][0][0] << std::endl;
+        //------------------------------------------------------------------------------
+        //------------------------------------------------------------------------------
+        //------------------------------------------------------------------------------
+        //
+        effTreeFD->Fill();
+        ND_OffAxis_Unrotated_Sim_hadronic_hit.clear();
+        ND_OffAxis_Sim_hadronic_hit.clear();
+     } // end Loop over ND_vtx_vx_vec
 
-      } // end Loop over ND_vtx_vx_vec
-
-      ND_Sim_hadron_contain_result_before_throw.emplace_back(hadron_contain_result_before_throw_vec_for_vtx_vx);
       ND_Sim_hadron_throw_result.emplace_back(hadron_throw_result_vec_for_vtx_vx);
 
     }   // end Loop over ND_off_axis_pos_vec
-
-    effTreeFD->Fill();
     iwritten++;
+    ND_RandomVtx_Sim_hadronic_hit.clear();
+    ND_OnAxis_Sim_hadronic_hit.clear();
 
-  }     // end loop over events
-
+  } // end loop over events entries
   std::cout << "Written evts: " << iwritten << std::endl;
-
+  myfile << "Written evts: " << iwritten << "\n";
+  //
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //------------------------------------------------------------------------------
+  //
   // Write trees
   TFile * outFile = new TFile("Output_FDGeoEff.root", "RECREATE");
   ThrowsFD->Write();
   effTreeFD->Write();
   hist_vetoEnergyFD->Write();
-  hist_vetoEnergyFD_new->Write();
 
+  myfile.close();
   outFile->Close();
-
 } // end main
